requirements.txt
requests==2.32.3
networkx==3.3
matplotlib==3.9.2
python-dotenv==1.0.1
pandas==2.2.2